parse logic inside get_partition_range func
