+++ metadata-ingestion/tests/unit/test_bq_get_partition_range.py
@@ -1,8 +1,6 @@
 import datetime
 
-from datahub.utilities.parsing_partition_util import (
-    get_partition_range_from_partition_id,
-)
+from datahub.ingestion.source.sql.bigquery import get_partition_range_from_partition_id
 
 
 def test_get_partition_range_from_partition_id():
@@ -16,6 +14,10 @@ def test_get_partition_range_from_partition_id():
     assert get_partition_range_from_partition_id(
         "2022", datetime.datetime(2021, 5, 2)
     ) == (datetime.datetime(2021, 1, 1), datetime.datetime(2022, 1, 1))
+    assert get_partition_range_from_partition_id("2022", None) == (
+        datetime.datetime(2022, 1, 1),
+        datetime.datetime(2023, 1, 1),
+    )
     # monthly partition check
     assert get_partition_range_from_partition_id(
         "202202", datetime.datetime(2022, 2, 1)
@@ -26,6 +28,10 @@ def test_get_partition_range_from_partition_id():
     assert get_partition_range_from_partition_id(
         "202202", datetime.datetime(2021, 12, 13)
     ) == (datetime.datetime(2021, 12, 1), datetime.datetime(2022, 1, 1))
+    assert get_partition_range_from_partition_id("202202", None) == (
+        datetime.datetime(2022, 2, 1),
+        datetime.datetime(2022, 3, 1),
+    )
     # daily partition check
     assert get_partition_range_from_partition_id(
         "20220205", datetime.datetime(2022, 2, 5)
@@ -33,6 +39,10 @@ def test_get_partition_range_from_partition_id():
     assert get_partition_range_from_partition_id(
         "20220205", datetime.datetime(2022, 2, 3)
     ) == (datetime.datetime(2022, 2, 3), datetime.datetime(2022, 2, 4))
+    assert get_partition_range_from_partition_id("20220205", None) == (
+        datetime.datetime(2022, 2, 5),
+        datetime.datetime(2022, 2, 6),
+    )
     # hourly partition check
     assert get_partition_range_from_partition_id(
         "2022020509", datetime.datetime(2022, 2, 5, 9)
@@ -40,3 +50,7 @@ def test_get_partition_range_from_partition_id():
     assert get_partition_range_from_partition_id(
         "2022020509", datetime.datetime(2022, 2, 3, 1)
     ) == (datetime.datetime(2022, 2, 3, 1), datetime.datetime(2022, 2, 3, 2))
+    assert get_partition_range_from_partition_id("2022020509", None) == (
+        datetime.datetime(2022, 2, 5, 9),
+        datetime.datetime(2022, 2, 5, 10),
+    )
